devops/aws/requirements.txt
boto3==1.17.1
botocore==1.20.1
PyJWT==2.4.0
requests==2.28.1
argparse==1.4.0